fix добавлены отдельные интерфейсы для хранилищ пользователей и фильмов, добавлены сервисные классы

+++ src/main/java/ru/yandex/practicum/filmorate/controller/Controller.java
@@ -7,22 +7,22 @@ import java.util.HashMap;
 import java.util.Map;
 
 abstract class Controller<T extends Entity> {
-    protected int identifier = 0;
-    protected final Map<Integer, T> storage = new HashMap<>();
+    protected long identifier = 0;
+    protected final Map<Long, T> storage = new HashMap<>();
 
     protected Collection<T> getAllValues() {
         return storage.values();
     }
 
-    protected int getIdentifier() {
+    protected long getIdentifier() {
         return ++identifier;
     }
 
-    public void setIdentifier(int identifier) {
+    public void setIdentifier(long identifier) {
         this.identifier = identifier;
     }
 
-    public Map<Integer, T> getStorage() {
+    public Map<Long, T> getStorage() {
         return storage;
     }
 }
